Rename activities reducer to activitiesReducer
